requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.6,<3.0
python-jose[cryptography]>=3.3,<4.0
bcrypt>=5.0,<6.0
reportlab>=4.2,<5.0
python-multipart>=0.0.9,<0.1
pytest>=8.3,<9.0
httpx>=0.27,<1.0
